src/requirements/base.txt
aioredis==1.3.1
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crossplane==0.5.3
cryptography==2.8
daphne==2.4.1
Django==2.2.11
django-autocomplete-light==3.5.1
django-cas-ng==4.1.1
django-celery==3.3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-guardian==2.2.0
django-js-asset==1.2.2
django-mama-cas==2.4.0
django-mdeditor==0.1.17
django-rest-swagger==2.2.0
django-spcc>=0.2.0
django-timezone-field==4.0
django-uwsgi==0.2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
drfdocs==0.0.11
flower==0.9.3
freeze==1.0.10
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.3.1
ipaddress==1.0.23
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==1.4.6
openapi-codec==1.3.2
packaging==20.3
paramiko==2.7.1
Pillow==7.0.0
prettytable==0.7.2
psutil==5.7.0
py-zabbix==1.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
pyecharts==1.7.0
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-cas==1.5.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
spmo>=2.0.9
sqlparse==0.3.1
swagger-spec-validator==2.5.0
tornado==5.1.1
Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
uwsgitop==0.11
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==4.7.1
